refactor: Update command and query models for mission assignments
- Changed id type to long
- Now we use primitives instead of value objects for the commands

+++ src/main/java/com/spire/platform/u202310680/missions/domain/model/queries/GetMissionAssignmentByIdQuery.java
@@ -1,4 +1,4 @@
 package com.spire.platform.u202310680.missions.domain.model.queries;
 
-public record GetMissionAssignmentByIdQuery(Integer id) {
+public record GetMissionAssignmentByIdQuery(Long id) {
 }
